requirements.txt
Django==1.10
Pillow==3.3.1
django-imagekit==3.3
django-extensions==1.7.3
numpy==1.11.1
django-crispy-forms==1.6.0
